# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mo AI Voice Assistant API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type ConversationMessageRole =
  (typeof ConversationMessageRole)[keyof typeof ConversationMessageRole];

export const ConversationMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ConversationMessage {
  role: ConversationMessageRole;
  content: string;
}

export type UserPreferencesResponseLength =
  (typeof UserPreferencesResponseLength)[keyof typeof UserPreferencesResponseLength];

export const UserPreferencesResponseLength = {
  short: "short",
  medium: "medium",
  long: "long",
} as const;

export interface UserPreferences {
  name?: string;
  location?: string;
  timezone?: string;
  responseLength?: UserPreferencesResponseLength;
}

export interface ReminderData {
  title: string;
  content: string;
  datetime: string;
}

export interface NoteData {
  content: string;
  timestamp?: string;
}

export type ChatRequestMode =
  (typeof ChatRequestMode)[keyof typeof ChatRequestMode];

export const ChatRequestMode = {
  executive: "executive",
  creative: "creative",
  motivational: "motivational",
  planner: "planner",
} as const;

export interface ChatRequest {
  message: string;
  mode?: ChatRequestMode;
  messages?: ConversationMessage[];
  preferences?: UserPreferences;
}

export interface ChatResponse {
  reply: string;
  functionCalled?: string;
  reminder?: ReminderData;
  note?: NoteData;
}

export interface SpeakRequest {
  text: string;
}

export type VoiceRequestFormat =
  (typeof VoiceRequestFormat)[keyof typeof VoiceRequestFormat];

export const VoiceRequestFormat = {
  m4a: "m4a",
  mp4: "mp4",
  wav: "wav",
  caf: "caf",
} as const;

export type VoiceRequestMode =
  (typeof VoiceRequestMode)[keyof typeof VoiceRequestMode];

export const VoiceRequestMode = {
  executive: "executive",
  creative: "creative",
  motivational: "motivational",
  planner: "planner",
} as const;

export interface VoiceRequest {
  /** Base64-encoded audio file */
  audio: string;
  format?: VoiceRequestFormat;
  mode?: VoiceRequestMode;
  /** Previous conversation turns for continuity */
  messages?: ConversationMessage[];
  preferences?: UserPreferences;
}

export interface VoiceResponse {
  transcript: string;
  reply: string;
  audioBase64: string;
  functionCalled?: string;
  reminder?: ReminderData;
  note?: NoteData;
}

export interface ErrorResponse {
  error: string;
}