# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Mo AI Voice Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: mo
    description: Mo AI voice assistant operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /mo/chat:
    post:
      operationId: moChat
      tags: [mo]
      summary: Send a message to Mo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Mo's response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mo/speak:
    post:
      operationId: moSpeak
      tags: [mo]
      summary: Convert text to speech via ElevenLabs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakRequest"
      responses:
        "200":
          description: Audio stream (mpeg)
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mo/voice:
    post:
      operationId: moVoice
      tags: [mo]
      summary: Full voice pipeline for native mobile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceRequest"
      responses:
        "200":
          description: Transcript, reply text, and base64 audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ConversationMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required: [role, content]
    UserPreferences:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        timezone:
          type: string
        responseLength:
          type: string
          enum: [short, medium, long]
          default: medium
    ReminderData:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        datetime:
          type: string
      required: [title, content, datetime]
    NoteData:
      type: object
      properties:
        content:
          type: string
        timestamp:
          type: string
      required: [content]
    ChatRequest:
      type: object
      properties:
        message:
          type: string
        mode:
          type: string
          enum: [executive, creative, motivational, planner]
          default: executive
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
        preferences:
          $ref: "#/components/schemas/UserPreferences"
      required: [message]
    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        functionCalled:
          type: string
        reminder:
          $ref: "#/components/schemas/ReminderData"
        note:
          $ref: "#/components/schemas/NoteData"
      required: [reply]
    SpeakRequest:
      type: object
      properties:
        text:
          type: string
      required: [text]
    VoiceRequest:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio file
        format:
          type: string
          enum: [m4a, mp4, wav, caf]
          default: m4a
        mode:
          type: string
          enum: [executive, creative, motivational, planner]
          default: executive
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
          description: Previous conversation turns for continuity
        preferences:
          $ref: "#/components/schemas/UserPreferences"
      required: [audio]
    VoiceResponse:
      type: object
      properties:
        transcript:
          type: string
        reply:
          type: string
        audioBase64:
          type: string
        functionCalled:
          type: string
        reminder:
          $ref: "#/components/schemas/ReminderData"
        note:
          $ref: "#/components/schemas/NoteData"
      required: [transcript, reply, audioBase64]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]