# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mo AI Voice Assistant API
 * OpenAPI spec version: 0.2.0
 */
import type { ChatRequestMode } from "./chatRequestMode";
import type { ConversationMessage } from "./conversationMessage";
import type { UserPreferences } from "./userPreferences";

export interface ChatRequest {
  message: string;
  mode?: ChatRequestMode;
  messages?: ConversationMessage[];
  preferences?: UserPreferences;
}