# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mo AI Voice Assistant API
 * OpenAPI spec version: 0.2.0
 */
import type { NoteData } from "./noteData";
import type { ReminderData } from "./reminderData";

export interface ChatResponse {
  reply: string;
  functionCalled?: string;
  reminder?: ReminderData;
  note?: NoteData;
}