# lib/api-zod/src/generated/types/conversationMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mo AI Voice Assistant API
 * OpenAPI spec version: 0.2.0
 */
import type { ConversationMessageRole } from "./conversationMessageRole";

export interface ConversationMessage {
  role: ConversationMessageRole;
  content: string;
}