# lib/api-zod/src/generated/types/noteData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mo AI Voice Assistant API
 * OpenAPI spec version: 0.2.0
 */

export interface NoteData {
  content: string;
  timestamp?: string;
}