# lib/api-zod/src/generated/types/userPreferences.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mo AI Voice Assistant API
 * OpenAPI spec version: 0.2.0
 */
import type { UserPreferencesResponseLength } from "./userPreferencesResponseLength";

export interface UserPreferences {
  name?: string;
  location?: string;
  timezone?: string;
  responseLength?: UserPreferencesResponseLength;
}